# src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CategoryDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"postalCode",
					"address",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"email",
					"postalCode",
					"address",
					"phone",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUser": {
				"properties": {
					"password": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LoginDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "refresh-bookstore",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/category": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"sessionAuth": [
							"isAdmin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCategoryList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"sessionAuth": [
							"isAdmin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"sessionAuth": [
							"isAdmin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/category/{id}": {
			"get": {
				"operationId": "GetCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/": {
			"get": {
				"operationId": "GetHomePage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [],
				"parameters": []
			}
		},
		"/login": {
			"get": {
				"operationId": "GetLoginPage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Login",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDTO"
							}
						}
					}
				}
			}
		},
		"/register": {
			"get": {
				"operationId": "GetRegisterPage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user-mypage": {
			"get": {
				"operationId": "GetUserMypage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			}
		},
		"/cart": {
			"get": {
				"operationId": "GetCart",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [],
				"parameters": []
			}
		},
		"/order-detail": {
			"get": {
				"operationId": "GetOrderDetail",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			}
		},
		"/order-create": {
			"get": {
				"operationId": "GetOrderCreate",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			}
		},
		"/order-complete": {
			"get": {
				"operationId": "GetOrderComplete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			}
		},
		"/order-list": {
			"get": {
				"operationId": "GetOrderList",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			}
		},
		"/book-detail": {
			"get": {
				"operationId": "GetBookDetail",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [],
				"parameters": []
			}
		},
		"/book-search": {
			"get": {
				"operationId": "GetBookSearch",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user-admin": {
			"get": {
				"operationId": "GetUserAdmin",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Page"
				],
				"security": [
					{
						"sessionAuth": [
							"isAdmin"
						]
					}
				],
				"parameters": []
			}
		},
		"/user": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUser"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUser"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDTO"
							}
						}
					}
				}
			}
		},
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"sessionAuth": [
							"isAdmin"
						]
					}
				],
				"parameters": []
			}
		},
		"/login/status": {
			"get": {
				"operationId": "CheckLoginStatus",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"sessionAuth": []
					}
				],
				"parameters": []
			}
		},
		"/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}